requirements.txt
Django==1.10

# This fork fixes Django 1.10 support.
https://github.com/davidwtbuxton/djangae/archive/use-argparse-for-runserver.zip

GoogleAppEngineCloudStorageClient==1.9.22.1
Pygments==2.1.3
